backend/requirements.txt
# Flask and Extensions
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0

# Database
SQLAlchemy==2.0.23

# Document Parsing
PyPDF2==3.0.1
python-docx==1.1.0

# NLP and ML
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2

# Security
bcrypt==4.1.2

# Utilities
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0
psycopg2-binary==2.9.9
